checked semantics are incorrect: must check let bindings
The checked semantics are not strict enough. For let bindings, they
substitute `e2[x := e2]` and check that, but this will miss checking
any unused bindings.
For example:

```
let _check = check false in 1
```

The above will succeed because `_check` is not mentioned.

This would be straightforward to fix by adding a `check e1` clause to
the checked semantics, but, this change means that the checked semantics
are no longer "straightforward derivations" of the substitution-based
bigstep semantics. If we depart from the bigstep semantics anyway, then
we could probably define a simpler semantics altogether.